+++ oya-back/app/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import auth, main, report
