+++ oya-back/app/controllers/report.py
@@ -6,18 +6,11 @@ from sqlalchemy import (
     select,
     case,
     literal,
-    literal_column,
     and_,
 )
 import datetime
-from typing import List
-from . import schemas
-from .models import Interval, Activity, Entry, Association
-from fastapi import APIRouter, Depends
-from .database import get_db
-from .auth import User as UserSchema, get_current_user
-
-router = APIRouter()
+from ..schemas.auth import User as UserSchema
+from ..models import Interval, Activity, Entry, Association
 
 
 def get_report_cte(
@@ -160,108 +153,3 @@ def get_report_cte(
     )
 
     return (sq, common_columns)
-
-
-@router.get("/totals/", tags=["Reports"], response_model=schemas.ReportBase)
-async def get_total_report(
-    from_date: datetime.date | datetime.datetime = None,
-    to_date: datetime.date | datetime.datetime = None,
-    tick: datetime.timedelta = datetime.timedelta(days=1),
-    db: Session = Depends(get_db),
-    user: UserSchema = Depends(get_current_user),
-):
-    (sq, common_columns) = get_report_cte(
-        db=db, from_date=from_date, to_date=to_date, tick=tick, user=user
-    )
-    stmt = select(*common_columns)
-    return db.execute(stmt).first()
-
-
-@router.get(
-    "/totals-periodic/", tags=["Reports"], response_model=List[schemas.ReportPeriodic]
-)
-def get_periodic_total_report(
-    from_date: datetime.date | datetime.datetime = None,
-    to_date: datetime.date | datetime.datetime = None,
-    tick: datetime.timedelta = datetime.timedelta(days=1),
-    db: Session = Depends(get_db),
-    user: UserSchema = Depends(get_current_user),
-):
-    (sq, common_columns) = get_report_cte(
-        db=db, from_date=from_date, to_date=to_date, tick=tick, user=user
-    )
-    stmt = (
-        select(sq.c.index, *common_columns)
-        .group_by(sq.c.index)
-        .order_by(sq.c.index.desc())
-    )
-    return db.execute(stmt).all()
-
-
-@router.get(
-    "/activities/", tags=["Reports"], response_model=List[schemas.ReportActivities]
-)
-def get_activities_report(
-    from_date: datetime.date | datetime.datetime = None,
-    to_date: datetime.date | datetime.datetime = None,
-    tick: datetime.timedelta = datetime.timedelta(days=1),
-    db: Session = Depends(get_db),
-    user: UserSchema = Depends(get_current_user),
-):
-    (sq, common_columns) = get_report_cte(
-        db=db, from_date=from_date, to_date=to_date, tick=tick, user=user
-    )
-    stmt = (
-        select(sq.c.parent_id.label("activity_id"), *common_columns)
-        .group_by(sq.c.parent_id)
-        .order_by(sq.c.parent_id)
-    )
-    return db.execute(stmt).all()
-
-
-@router.get(
-    "/activities-periodic/",
-    tags=["Reports"],
-    response_model=List[schemas.ReportActivitiesPeriodic],
-)
-def get_periodic_activities_report(
-    from_date: datetime.date | datetime.datetime = None,
-    to_date: datetime.date | datetime.datetime = None,
-    tick: datetime.timedelta = datetime.timedelta(days=1),
-    db: Session = Depends(get_db),
-    user: UserSchema = Depends(get_current_user),
-):
-    (sq, common_columns) = get_report_cte(
-        db=db, from_date=from_date, to_date=to_date, tick=tick, user=user
-    )
-    stmt = (
-        select(sq.c.index, sq.c.parent_id.label("activity_id"), *common_columns)
-        .group_by(sq.c.index, sq.c.parent_id)
-        .order_by(sq.c.index.desc(), sq.c.parent_id)
-    )
-    return db.execute(stmt).all()
-
-
-@router.get(
-    "/activities-periodic/{activitiy_id}/",
-    tags=["Reports"],
-    response_model=List[schemas.ReportActivitiesPeriodic],
-)
-def get_periodic_activities_report(
-    activitiy_id: int,
-    from_date: datetime.date | datetime.datetime = None,
-    to_date: datetime.date | datetime.datetime = None,
-    tick: datetime.timedelta = datetime.timedelta(days=1),
-    db: Session = Depends(get_db),
-    user: UserSchema = Depends(get_current_user),
-):
-    (sq, common_columns) = get_report_cte(
-        db=db, from_date=from_date, to_date=to_date, tick=tick, user=user
-    )
-    stmt = (
-        select(sq.c.index, sq.c.parent_id.label("activity_id"), *common_columns)
-        .where(sq.c.parent_id == activitiy_id)
-        .group_by(sq.c.index, sq.c.parent_id)
-        .order_by(sq.c.index.desc(), sq.c.parent_id)
-    )
-    return db.execute(stmt).all()
