+++ oya-back/app/routers/__init__.py
@@ -1,0 +1,3 @@
+from .auth import router as auth
+from .main import router as main
+from .report import router as report
