+++ oya-back/app/schemas/__init__.py
@@ -1,0 +1,1 @@
+from . import main, auth, report
