+++ oya-back/app/schemas/auth.py
@@ -1,0 +1,37 @@
+from pydantic import BaseModel
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username: str | None = None
+
+
+class UserBase(BaseModel):
+    username: str
+    email: str | None = None
+    firstname: str | None = None
+    lastname: str | None = None
+    superuser: bool | None = None
+
+    class Config:
+        orm_mode = True
+
+
+class User(UserBase):
+    id: int
+
+
+class UserCreate(UserBase):
+    password: str
+
+
+class UserUpdate(UserBase):
+    username: str | None
+    password: str | None
+
+
+class UserInDB(User):
+    password: str
