+++ oya-back/app/schemas/main.py
@@ -89,62 +89,5 @@ class IntervalsResponse(BaseModel):
     data: List[Interval]
 
 
-class ReportSinge(BaseModel):
-    duration: int
-    occurance: int
-
-    class Config:
-        orm_mode = True
-
-
-class Period(BaseModel):
-    s: str
-    e: str
-    d: int
-    note: Optional[str]
-    entries: List[Entry]
-
-    class Config:
-        orm_mode = True
-
-
-class DailyReport(BaseModel):
-    date: str
-    periods: List[Period]
-    report: Dict[str, ReportSinge]
-
-    class Config:
-        orm_mode = True
-
-
-class ReportBase(BaseModel):
-    time: datetime.timedelta
-    time_pure: datetime.timedelta
-    occurance_in_period: int
-    occurance: int
-    occurance_pure: int
-    periods: int
-    all_periods: int
-    intervals_count: int
-    intervals_total_time: datetime.timedelta
-    intervals_total_time_pure: datetime.timedelta
-    period_start: datetime.datetime
-    period_end: datetime.datetime
-    period_range: datetime.timedelta
-
-
-class ReportPeriodic(ReportBase):
-    index: int
-
-
-class ReportActivities(ReportBase):
-    activity_id: int
-
-
-class ReportActivitiesPeriodic(ReportBase):
-    index: int
-    activity_id: int
-
-
 Activity.update_forward_refs()
 Interval.update_forward_refs()
